requirements.txt
# Frontend
streamlit==1.41.1

# LLM & RAG pipeline
langchain==0.3.16
langchain-community==0.3.16
langchain-core>=0.3.30
langchain-text-splitters>=0.3.0
ollama>=0.1.7

# Vector search & Embedding
faiss-cpu>=1.9.0
sentence-transformers>=2.6.0
transformers>=4.38,<5
accelerate>=0.25.0

# PDF / DOC processing
pymupdf==1.24.1
rapidocr-onnxruntime
pillow
pdfplumber==0.10.0
python-docx>=1.0.0
docx2txt

# Scientific computing
numpy>=1.26.0
pandas>=2.2.0
torch>=2.2.0

# Utilities
tqdm>=4.65.0